requirements.txt
pyhdf
numpy
matplotlib
torch
scikit-learn
pandas
scipy
earthaccess
pyyaml
tqdm
imageio
Pillow
python-pptx
tabpfn
